[BEAM-4164] Move former CassandraServiceImplTest utils tests to CassandraIOTest
